+++ Wire-iOS/Sources/UserInterface/StartUI/StartUI/StartUIViewController.swift
@@ -170,6 +170,31 @@ final class StartUIViewController: UIViewController, SpinnerCapable {
         view.accessibilityViewIsModal = true
     }
 
+    private func createConstraints() {
+        [searchHeaderViewController.view, groupSelector, searchResultsViewController.view].forEach { $0?.translatesAutoresizingMaskIntoConstraints = false }
+
+        searchHeaderViewController.view.fitInSuperview(exclude: [.bottom])
+
+        if showsGroupSelector {
+            NSLayoutConstraint.activate([
+                groupSelector.topAnchor.constraint(equalTo: searchHeaderViewController.view.bottomAnchor),
+                searchResultsViewController.view.topAnchor.constraint(equalTo: groupSelector.bottomAnchor)
+                ])
+
+            groupSelector.fitInSuperview(exclude: [.bottom, .top])
+        } else {
+            NSLayoutConstraint.activate([
+            searchResultsViewController.view.topAnchor.constraint(equalTo: searchHeaderViewController.view.bottomAnchor)
+                ])
+        }
+
+        searchResultsViewController.view.fitInSuperview(exclude: [.top])
+    }
+
+    var showsGroupSelector: Bool {
+        return SearchGroup.all.count > 1 && ZMUser.selfUser().canSeeServices
+    }
+
     func showKeyboardIfNeeded() {
         let conversationCount = ZMConversationList.conversations(inUserSession: ZMUserSession.shared()!).count /// TODO: unwrap
         if conversationCount > StartUIViewController.InitiallyShowsKeyboardConversationThreshold {
